refactor(core): refactor debug_node to move classes/query into separate files
